# server/openapi.yaml
openapi: 3.1.0
info:
  title: UCT ERP Secured API Portal
  description: |
    Production-grade, enterprise-ready API endpoints for Universidade Católica Timorense (UCT) ERP.
    
    ## Authentication
    All endpoints (except `/api/auth/login`) require authentication via JWT Bearer Token.
    Provide the token in the headers as:
    `Authorization: Bearer <JWT_TOKEN>`

    ## Role-Based Access Control (RBAC)
    Operations and visibility are restricted by portal role:
    - **Administrator**: Unrestricted read/write/delete access.
    - **Student**: Access restricted to own student details, attendance, and fee transactions.
    - **Staff/Faculty**: Access to manage students (read/write) and mark attendance.
    - **HR**: Access to manage employee records and payroll.
    - **Accounts**: Access to manage transactions, ledgers, and fee structures.
    - **Transport**: Access to manage transport routes, vehicles, and allocations.
  version: 2.0.0
servers:
  - url: http://localhost:5000
    description: Local Express Backend Server
paths:
...
  /api/students:
    get:
      summary: Get Student Records
      description: |
        Returns student records.
        - **Student role**: Returns only the student's own record matching their logged-in scholar number.
        - **Staff/Faculty, Administrator**: Returns all student records.
        - **Other roles**: Returns 403 Forbidden.
      tags:
        - Student Module
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paginated student details list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '403':
          description: Insufficient permissions

    post:
      summary: Create New Student Record
      description: Adds a new student to the database (Restricted to Administrator and Staff/Faculty).
      tags:
        - Student Module
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '201':
          description: Student created successfully
        '400':
          description: Validation error or unique constraints violated (Scholar No already exists)
        '403':
          description: Insufficient permissions
  /api/students/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          description: Database record ID
    put:
      summary: Update Student Record
      description: Modifies student details (Restricted to Administrator and Staff/Faculty).
      tags:
        - Student Module
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: Student updated successfully
        '403':
          description: Insufficient permissions
        '404':
          description: Student record not found

    delete:
      summary: Delete Student Record
      description: Permanently deletes a student record (Restricted to Administrator only).
      tags:
        - Student Module
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student deleted successfully
        '403':
          description: Insufficient permissions (Non-admin)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Standard Authorization Bearer token header.
  schemas:
    Student:
      type: object
      required:
        - scholar_no
        - name
        - course
        - branch
        - batch
      properties:
        id:
          type: integer
          readOnly: true
        scholar_no:
          type: string
        enroll_no:
          type: string
        roll_no:
          type: string
        name:
          type: string
        course:
          type: string
        branch:
          type: string
        batch:
          type: string
        category:
          type: string
        admission_date:
          type: string
        status:
          type: string
        due_fee:
          type: number
        json_details:
          type: string
          description: Stringified JSON containing custom form fields
...